+++ lib/src/services/config_repository.dart
@@ -20,6 +20,26 @@ class ConfigRepository {
     config.save(_configPath);
   }
 
+  static void update({
+    String? cachePath,
+    bool? useGitCache,
+    String? gitCachePath,
+    String? flutterUrl,
+    bool? disableUpdateCheck,
+    DateTime? lastUpdateCheck,
+  }) {
+    final currentConfig = loadFile();
+    final newConfig = currentConfig.copyWith(
+      cachePath: cachePath,
+      useGitCache: useGitCache,
+      gitCachePath: gitCachePath,
+      flutterUrl: flutterUrl,
+      disableUpdateCheck: disableUpdateCheck,
+      lastUpdateCheck: lastUpdateCheck,
+    );
+    save(newConfig);
+  }
+
   static Config loadEnv() {
     final environments = Platform.environment;
 
@@ -30,12 +50,14 @@ class ConfigRepository {
 
     for (final variable in ConfigKeys.values) {
       final value = environments[variable.envKey];
-      if (value == null) continue;
+      final legacyFvmHome = environments['FVM_HOME'];
 
       if (variable == ConfigKeys.cachePath) {
-        cachePath = value;
+        cachePath = value ?? legacyFvmHome;
         break;
       }
+
+      if (value == null) continue;
 
       if (variable == ConfigKeys.useGitCache) {
         gitCache = value == 'true';